% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_bf.R
\name{download_bf}
\alias{download_bf}
\title{Download Bolsa Família}
\usage{
download_bf(year = NULL, month = NULL, query = "payments", ...)
}
\arguments{
\item{year}{\code{integer} between 2013 and 2018}

\item{month}{\code{integer} between 1 and 12}

\item{query}{a \code{character} string indicating whether to download payment ('payment' or 'pagamento') data or withdrawal ('withdrawal' or 'saque') data}

\item{...}{additional parameters}
}
\description{
Download of the Bolsa Família payment and withdrawal data.
}
\concept{Benefícios ao cidadão}
