% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_compras}
\alias{download_compras}
\title{Download Compras}
\usage{
download_compras(year = NULL, month = NULL, interactive = FALSE,
  file = NULL, ...)
}
\arguments{
\item{year}{\code{integer} between 2014 and 2019}

\item{month}{\code{integer} between 1 and 12}

\item{interactive}{a \code{boolean}. If \code{TRUE}, then
\code{file} will be ignored.}

\item{file}{\code{numeric} Which file do you want to access? The options are:
1 for Apostilamento, 2 for Compras, 3 for Item, 4 for Termo Aditivo or
5 for a merged file using all these information. This parameter
will be used only if \code{interactive = FALSE}.}

\item{...}{additional parameters}
}
\value{
\code{data.frame}
}
\description{
This functions is not ready for use.
}
