% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_cp}
\alias{download_cp}
\title{Download Cartoes de Pagamentos}
\usage{
download_cp(ym = NULL, type = NULL, ...)
}
\arguments{
\item{ym}{\code{character} year-month, see examples.}

\item{type}{must be 'cpgf' (cartao de pagamentos do governo federal),
'cpcc' (cartao de pagamentos do governo federal - compras centralizadas)
or 'cpdc' (cartao de pagamentos da defesa civil)}

\item{...}{additional parameters}
}
\value{
a \code{data.table} containing the data
}
\description{
Download Cartoes de Pagamentos
}
\examples{

(x <- download_cp(ym = "201401", type = 'cpgf'))
(x <- download_cp(ym = "201501", type = 'cpcc'))
(x <- download_cp(ym = c("201501", "201502"), type = 'cpdc'))


}
\concept{Cartao de Pagamento}
