% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_exec_desp}
\alias{download_exec_desp}
\title{Download Execucao de Despesas}
\usage{
download_exec_desp(ym = NULL, ...)
}
\arguments{
\item{ym}{\code{character} year-month, see examples.}

\item{...}{additional parameters}
}
\value{
\code{data.table}
}
\description{
Download Execucao de Despesas
}
\examples{

(x <- download_exec_desp(ym = "201401"))
(x <- download_exec_desp(ym = c("201401", "201501")))
(x <- download_exec_desp(ym = c("201501", "201502")))

\dontrun{
 (x <- download_exec_desp(year = "2014", month = 2))
 (x <- download_exec_desp(year = 2014))
}

}
\seealso{
Other Despesas Publicas: 
\code{\link{download_transf}()}
}
\concept{Despesas Publicas}
