% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rewrite.R
\name{download_lic}
\alias{download_lic}
\title{Download Licitacoes}
\usage{
download_lic(year = NULL, month = NULL, interactive = NULL,
  file = 3, ...)
}
\arguments{
\item{year}{\code{integer} between 2014 and 2018}

\item{month}{\code{integer} between 1 and 12}

\item{interactive}{a \code{boolean}. If \code{TRUE}, then
\code{file} will be ignored.}

\item{file}{\code{numeric} Which file do you want to access? The options are:
1 for Pagamento, 2 for Passagem, 3 for Trecho, 4 for Viagem or
5 for a merged file using all these information. This parameter
will only be used if \code{interactive = FALSE}.}

\item{...}{additional parameters}
}
\value{
\code{data.frame}
}
\description{
This functions is not ready for use.
}
