% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_lic_cont}
\alias{download_lic_cont}
\title{Download Licitacoes e Compras}
\usage{
download_lic_cont(year = NULL, month = NULL, type = "licitacoes", ...)
}
\arguments{
\item{year}{\code{integer} between 2014 and 2018}

\item{month}{\code{integer} between 1 and 12}

\item{type}{\code{character} must be 'licitacoes' or 'compras'}

\item{...}{additional parameters}
}
\value{
\code{data.frame}
}
\description{
Download Licitacoes e Compras
}
