% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_orcamento}
\alias{download_orcamento}
\title{Download Orcamento}
\usage{
download_orcamento(year = NULL, ...)
}
\arguments{
\item{year}{\code{integer} between 2014 and 2018}

\item{...}{additional parameters}
}
\value{
\code{data.frame}
}
\description{
Download Orcamento
}
\examples{

(x <- download_orcamento(year = 2014))
(x <- download_orcamento(year = 2014:2015))

\dontrun{
 (x <- download_orcamento(year = "2014"))
 (x <- download_orcamento(year = 2014, month = 2))
}

}
\concept{Orcamento Publico}
