% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_receitas}
\alias{download_receitas}
\title{Download Receitas}
\usage{
download_receitas(year = NULL, ...)
}
\arguments{
\item{year}{\code{integer} between 2013 and 2018}

\item{...}{additional parameters}
}
\value{
\code{data.frame}
}
\description{
Download Receitas
}
