% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_transf}
\alias{download_transf}
\title{Download Transferencia}
\usage{
download_transf(ym = NULL, ...)
}
\arguments{
\item{ym}{\code{character} year-month, see examples.}

\item{...}{additional parameters}
}
\value{
a \code{data.table} containing the data
}
\description{
Download Transferencia
}
\examples{

(x <- download_transf(ym = "201401"))
(x <- download_transf(ym = c("201401", "201501")))
(x <- download_transf(ym = c("201501", "201502")))

\dontrun{
 (x <- download_transf(year = 2014))
}

}
\seealso{
Other Despesas Publicas: 
\code{\link{download_exec_desp}()}
}
\concept{Despesas Publicas}
