% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_viagens}
\alias{download_viagens}
\title{Download Viagens}
\usage{
download_viagens(year = NULL, interactive = interactive(), file = 5,
  ...)
}
\arguments{
\item{year}{\code{integer} between 2013 and 2018}

\item{interactive}{a \code{boolean}. If \code{TRUE}, then
\code{file} will be ignored.}

\item{file}{Which file do you want to access? The options are:
1 for Pagamento, 2 for Passagem, 3 for Trecho, 4 for Viagem or
5 for a merged file using all these information. This parameter
will only be used if \code{interactive = FALSE}.}

\item{...}{additional parameters}
}
\value{
\code{data.table}
}
\description{
Download Viagens
}
\examples{

\dontrun{
 (x <- download_viagens(year = 2018))
 (x <- download_viagens(year = 2017:2018))

 (x <- download_viagens(year = 2014, interactive = F, file = 3))
}

}
\concept{Viagens}
