% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fix_name}
\alias{fix_name}
\title{Fix colnames}
\usage{
fix_name(x)
}
\arguments{
\item{x}{\code{colnaes(df)}}
}
\value{
a \code{string} vector with same length than \code{x}
}
\description{
Auxiliar function to standardize variables' names.
}
