% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_cp}
\alias{download_cp}
\title{Download Cartoes de Pagamentos}
\usage{
download_cp(year = NULL, month = NULL, type = NULL, ...)
}
\arguments{
\item{year}{\code{integer} between 2013 and 2018}

\item{month}{\code{integer} between 1 and 12}

\item{type}{must be 'cpgf' (cartao de pagamentos do governo federal),
'cpcc' (cartao de pagamentos do governo federal - compras centralizadas)
or 'cpdc' (cartao de pagamentos da defesa civil)}

\item{...}{additional parameters}
}
\value{
\code{data.frame}
}
\description{
Download Cartoes de Pagamentos
}
