% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_exec_desp}
\alias{download_exec_desp}
\title{Download Execucao de Despesas}
\usage{
download_exec_desp(year = NULL, month = NULL, ...)
}
\arguments{
\item{year}{\code{integer} between 2013 and 2019}

\item{month}{\code{integer} between 1 and 12}

\item{...}{additional parameters}
}
\value{
\code{data.frame}
}
\description{
Download Execucao de Despesas
}
\examples{

(x <- download_exec_desp(year = 2014, month = 1))
(x <- download_exec_desp(year = 2014:2015, month = 1))
(x <- download_exec_desp(year = 2015, month = 1:2))

\dontrun{
 (x <- download_exec_desp(year = "2014", month = 2))
 (x <- download_exec_desp(year = 2014))
}

}
\seealso{
Other Despesas Publicas: \code{\link{download_transf}}
}
\concept{Despesas Publicas}
