% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_receitas}
\alias{download_receitas}
\title{Download Receitas}
\usage{
download_receitas(year = NULL, ...)
}
\arguments{
\item{year}{\code{integer} between 2013 and 2019}

\item{...}{additional parameters}
}
\value{
\code{data.frame}
}
\description{
Download Receitas
}
\examples{

(x <- download_receitas(year = 2014))
(x <- download_receitas(year = 2014:2015))

\dontrun{
 (x <- download_receitas(year = "2014"))
 (x <- download_receitas(year = 2014, month = 2))
}

}
\concept{Receitas Publicas}
