% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_transf}
\alias{download_transf}
\title{Download Transferencia}
\usage{
download_transf(year = NULL, month = NULL, ...)
}
\arguments{
\item{year}{\code{integer} between 2013 and 2018}

\item{month}{\code{integer} between 1 and 12}

\item{...}{additional parameters}
}
\value{
\code{data.frame}
}
\description{
Download Transferencia
}
