% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_transf}
\alias{download_transf}
\title{Download Transferencia}
\usage{
download_transf(year = NULL, month = NULL, ...)
}
\arguments{
\item{year}{\code{integer} between 2013 and 2018}

\item{month}{\code{integer} between 1 and 12}

\item{...}{additional parameters}
}
\value{
\code{data.frame}
}
\description{
Download Transferencia
}
\examples{

(x <- download_transf(year = 2014, month = 1))
(x <- download_transf(year = 2014:2015, month = 1))
(x <- download_transf(year = 2015, month = 1:2))

\dontrun{
 (x <- download_transf(year = "2014", month = 2))
 (x <- download_transf(year = 2014))
}

}
\seealso{
Other Despesas Publicas: \code{\link{download_exec_desp}}
}
\concept{Despesas Publicas}
